+++ server/controllers/api_controller.js
@@ -14,27 +14,28 @@ import { mime2ext } from '../../shared/modules/mime2ext';
 class ApiController {
 
   // Returns the proper file path if the file exists
-  static getFilePath(filename) {
-    const filePath = path.join(__dirname, `../uploads/${filename}`);
+  static getFilePath(key, filename) {
+    const filePath = path.join(__dirname, `../uploads/${key}/${filename}`);
     return (fs.existsSync(filePath)) ? filePath : null;
   }
 
-  constructor(db) { 
+  constructor(db) {
     this.db = db;
   }
-  
+
   async handleGet(key, res, options = {}) {
     const data = await this.db.models.entries.getEntry(key);
 
     // Return the data object if there's text available or if it's a file paste
     // and the file still exists
-    const stillExists = data &&
-      (data.text || (data.filename && ApiController.getFilePath(data.filename)));
+    const stillExists = data && (data.text
+      || (data.filename && ApiController.getFilePath(key, data.filename)));
 
     // Send back data the paste was forked from
     if (stillExists && data.forkedKey) {
       const forkedData = await this.db.models.entries.getEntry(data.forkedKey);
-      if (forkedData && forkedData.text && forkedData.privacy != privacyOptions.encrypted) {
+      if (forkedData && forkedData.text
+        && forkedData.privacy !== privacyOptions.encrypted) {
         data.forkedText = forkedData.text;
       } else {
         // Send back no forked key
@@ -56,7 +57,7 @@ class ApiController {
     const data = await this.db.models.entries.getEntry(key);
 
     // Don't return raw docs that are encrypted
-    if (data && data.text && data.privacy != privacyOptions.encrypted) {
+    if (data && data.text && data.privacy !== privacyOptions.encrypted) {
       winston.verbose('Retrieved raw document', { key });
       res.writeHead(200, contentType.plain);
       res.end(data.text);
@@ -69,14 +70,15 @@ class ApiController {
   }
 
   /**
-   * @param key
+   * @param params including key and filename
    * @param res
    */
-  async handleGetFile(key, res) {
-    const data = await this.db.models.entries.getEntry(key);
+  async handleGetFile(params, res) {
+    const data = await this.db.models.entries.getEntry(params.key);
 
     if (data && data.filename) {
-      const filePath = path.join(__dirname, `../uploads/${data.filename}`);
+      const filePath =
+        path.join(__dirname, `../uploads/${params.key}/${data.filename}`);
       if (fs.existsSync(filePath)) {
         return res.sendFile(filePath);
       }
@@ -84,7 +86,7 @@ class ApiController {
 
     fail(res, {
       clientMsg: 'File not found.',
-      serverMsg: 'Problem serving file: ' + key,
+      serverMsg: `Problem serving file: ${params.key}/${params.filename}`
     });
   }
 
@@ -103,10 +105,10 @@ class ApiController {
    */
   async handlePost(req, res) {
     const data = req.body,
-          isCurl = req.headers['user-agent'].includes('curl'),
-          host = req.headers.host;
+      isCurl = req.headers['user-agent'].includes('curl'),
+      host = req.headers.host;
 
-    data.file = (req.files) ?  req.files[0] : null;
+    data.file = (req.files) ? req.files[0] : null;
 
     // Do some handling here in case of posting using curl
     if (isCurl) {
@@ -130,10 +132,20 @@ class ApiController {
     // Try writing the file if it exists
     try {
       if (data.file) {
-        const ext = (data.file.originalname.match(/\..{1,20}$/) || [])[0]
-          || '';
-        data.filename = data.key + ext;
-        fs.writeFileSync(`server/uploads/${data.filename}`, data.file.buffer);
+        // We're guessing this is a pasted image, guess the file extension
+        if (data.file.originalname === 'file') {
+          data.filename = data.key + mime2ext[data.file.mimetype];
+        } else {
+          // Otherwise, keep the original file name up to 100 chars
+          data.filename = data.file.originalname.replace(/(^.{1,100})/, '$1');
+        }
+        // If there's no title, set the title to the filename
+        data.title = data.title || data.filename;
+
+        // Place the file in a directory by key because of duplicate filenames
+        fs.mkdirSync(`server/uploads/${data.key}`);
+        fs.writeFileSync(`server/uploads/${data.key}/${data.filename}`,
+          data.file.buffer);
       }
     } catch(err) {
       return fail(res, {
@@ -166,7 +178,7 @@ class ApiController {
     // Fail if we make it here
     fail(res, {
       clientMsg: 'Problem adding document. Please try again later.',
-      serverMsg: 'Error adding document for key: ' + data.key,
+      serverMsg: `Error adding document for key: ${data.key}`,
       resCode: 500
     });